feat: change create page to modal
