+++ frontend/src/admin/pages/[groupId]/workbook/[id]/index.vue
@@ -1,11 +1,19 @@
 <script setup lang="ts">
 import Button from '@/common/components/Atom/Button.vue'
+import InputItem from '@/common/components/Atom/InputItem.vue'
+import Modal from '@/common/components/Molecule/Modal.vue'
+import Switch from '@/common/components/Molecule/Switch.vue'
 import PaginationTable from '@/common/components/Organism/PaginationTable.vue'
+import TextEditor from '@/common/components/Organism/TextEditor.vue'
+import { NDatePicker } from 'naive-ui'
+import { ref } from 'vue'
 import IconTrash from '~icons/fa/trash-o'
 
-const props = defineProps<{
-  groupId: string
-}>()
+const show = ref(false)
+const title = ref('')
+const dateRange = ref<[number, number]>([1183135260000, Date.now()])
+const problemVisibleStat = ref(false)
+const difficultyVisibleStat = ref(false)
 </script>
 
 <template>
@@ -23,7 +31,7 @@ const props = defineProps<{
         <div>Total Problem: 10, Total Score: 100</div>
       </div>
       <div class="flex">
-        <Button class="mr-8 h-8">+ Create</Button>
+        <Button class="mr-8 h-8" @click="show = true">+ Create</Button>
         <Button class="h-8">Import</Button>
       </div>
     </div>
@@ -80,12 +88,6 @@ const props = defineProps<{
       :number-of-pages="3"
       :no-search-bar="true"
       :no-pagination="true"
-      @row-clicked="
-        (data) =>
-          $router.push(
-            `/admin/${props.groupId}/workbook/` + data.id + '/create'
-          )
-      "
     >
       <template #delete="{}">
         <Button class="flex h-[32px] w-[32px] items-center justify-center">
@@ -93,7 +95,7 @@ const props = defineProps<{
         </Button>
       </template>
     </PaginationTable>
-    <h2 class="mt-16 text-xl">Submission List</h2>
+    <h2 class="mt-10 text-xl">Submission List</h2>
     <PaginationTable
       :fields="[
         {
@@ -160,6 +162,50 @@ const props = defineProps<{
       </template>
     </PaginationTable>
   </div>
+  <Modal v-model="show" class="m-8 h-[96%] w-full px-40">
+    <div class="overflow-auto">
+      <div class="flex flex-col">
+        <div class="mt-10">
+          <h1 class="text-gray-dark mr-6 inline text-2xl font-semibold">
+            Create Workbook
+          </h1>
+        </div>
+        <div class="bg-gray h-[1px]"></div>
+        <div class="mt-10 flex">
+          <h2 class="mr-10 text-lg">Group</h2>
+          <h2 class="text-lg">Title</h2>
+        </div>
+        <div class="mt-6 flex">
+          <div class="mr-10">NPC 중급반</div>
+          <InputItem v-model="title" required placeholder="Title" />
+        </div>
+        <h2 class="mt-8 text-lg">Description</h2>
+        <TextEditor class="h-[360px] w-full" />
+
+        <div class="mt-10">
+          <NDatePicker
+            v-model:value="dateRange"
+            type="datetimerange"
+            clearable
+          />
+        </div>
+        <div class="mt-10 flex">
+          <Switch v-model="problemVisibleStat" class="mr-8" label="Visible" />
+          <Switch
+            v-model="difficultyVisibleStat"
+            label="Problem Difficulty Visible"
+          />
+        </div>
+        <div class="bg-gray mt-10 h-[1px]"></div>
+        <div class="mt-8 flex justify-end">
+          <div class="flex">
+            <Button class="mr-8 px-4 py-1" color="gray-dark">Cancel</Button>
+            <Button class="px-4 py-1">Next</Button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </Modal>
 </template>
 
 <route lang="yaml">
